♻️  Re-organize into workspace & make dotenv optional
* In order to make it easier to hack on other applications that re-use
some functionality (e.g. an Alexa skill), re-organize the source files
into:
  * A library crate containing the helpers to query the weather and pick
an outfit
  * A binary crate for the CLI tool that prints out an outfit. This was
just main.rs
* Added a feature flag `dotenv_key` to conditionally compile the
already-implemented ability to provide an OpenWeatherMap API key via an
environment variable specified by a .env file. This allows dotenv to be
just an optional dependency.
* Committed Cargo.lock to the repo since this is a binary crate
