+++ .github/workflows/rust.yml
@@ -31,5 +31,7 @@ jobs:
         token: ${{ secrets.GITHUB_TOKEN }}
     - name: Build
       run: cargo build --verbose
+    - name: Build with all features
+      run: cargo build --verbose --all-features
     - name: Run tests
       run: cargo test --verbose
