+++ picker/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "picker"
+description = "Common functions to get the weather and pick an outfit"
+version = "0.1.0"
+authors = ["Kesavan Yogeswaran <kesyog@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0"
+openweather = { git = "https://github.com/caemor/openweather"}
