+++ picker/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod gear;
+pub mod inputs;
+pub mod weather;
