+++ picker/src/weather.rs
